Rename decoder folder to faultTypes
